Authenticate GitHub access in actions workflows
In particular, these workflows install metamist-infrastructure via pip
by checking out the sample-metadata repo. Being authenticated means that
this works even when that repo is, for example, temporarily made private
for operational reasons.

Requires a suitable PAT to be set up and a cpg-infrastructure secret
referring to it, which should be three lines as follows:

    url=https://github.com/
    username=<username>
    password=<token>

The username is not used, so could be a placeholder like 'git'
or could be the token maker's username for documentary purposes.