Authenticate GitHub access in actions workflows (#147)
* Authenticate GitHub access in actions workflows

In particular, these workflows install metamist-infrastructure via pip
by checking out the sample-metadata repo. Being authenticated means that
this works even when that repo is, for example, temporarily made private
for operational reasons.

Requires a suitable PAT to be set up and a cpg-infrastructure secret
referring to it, which should be three lines as follows:

    url=https://github.com/
    username=<username>
    password=<token>

The username is not used, so could be a placeholder like 'git'
or could be the token maker's username for documentary purposes.

* Use current GitHub Actions versions (and avoid ::set-output)

Prevent "The following action(s) uses node12 which is deprecated"
warnings by updating to the current releases of the Actions used.

Rewrite `::set-output` as a write to $GITHUB_OUTPUT instead;
the `::save-state` and `::set-output` commands are also deprecated.