Change to storage visualisation function. (#305)
This commit is addressing the storage visualisation issue:

* Increase treemap job to 14GB, even hail is already reporting it has 14GB, so might not be needed.
* Limit error message to just 250 chars, this will prevent the spamming slack channel with large error dumps. 250 is just a random number.
* After testing the new RC of kaleido, generating image for all 3 levels is still freezing when tested locally.
As one of the possible solution was to use only 2 levels for the image and all 3 levels for the html page. So if people need more details they can follow the link from the Slack message. To avoid double parsing the data and creating 2 copies of rows, this PR is just filtering out dataframe rows by names. Field name contains full path to the stored object, so it just counts number of forward slashes.