Decode JSON string label values such as 'sequencing_groups'
Job attributes are transmitted as strings, including 'sequencing_groups'
which would naturally be a Python list of sequencing group IDs but is
encoded as a JSON list for transmission.

Decode the JSON string values of specified JSON_LABELS into lists etc.
When the labels are re-encoded as a dict in JSON string form, these
attribute values won't be double-encoded.