Remove deprecated skip_delete config in favour of new deletion_policy (#328)
Pulumi has deprecated the skip_delete config in favour of specifying a deletion policy. Having the deprecated skip_delete in our config was causing the deploy logs to be polluted with a message like:

```
gcp:organizations:Project {project}-gcp-project warning: urn:pulumi:production::datasets::gcp:organizations/project:Project::{project}-gcp-project verification warning: skip_delete is deprecated and will be removed in 6.0.0. Please use deletion_policy instead. A skip_delete value of false can be changed to a deletion_policy value of DELETE and a skip_delete value of true to a deletion_policy value of ABANDON for equivalent behavior.
```

for every project.

The equivalent of `skip_delete=True` is `deletion_policy='ABANDON'` so this change updates to that.